[SYCL] Do not use cl_* types in some tests
cl_* types are defined by SYCL spec as interoperability interface with
OpenCL. There is no need to use them in most of SYCL tests.

This change is mostly motivated by the fact that those cl_* type aliases
were moved from `sycl` into `sycl::opencl` namespace: we simply want to
avoid tests breakage when we align the implementation with the spec.